% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoes-package.R
\docType{package}
\name{volcanoes-package}
\alias{volcanoes-package}
\title{volcanoes: Holocene Volcanoes, Eruptions, and Eruption Events}
\description{
Tidy snapshots of the Smithsonian Institution Global Volcanism Program's Volcanoes of the World database. Provides three data frames covering Holocene volcanoes, their eruptions, and the events recorded during those eruptions. The schema follows the layout popularized by the rfordatascience/tidytuesday 2020-05-12 release, refreshed against the current Global Volcanism Program database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://volcano.si.edu/}
}

}
\author{
\strong{Maintainer}: Chester Ismay \email{chester.ismay@gmail.com}

}
\keyword{internal}
